+++ src/EntityGraphQL/Schema/MutationType.cs
@@ -26,9 +26,9 @@ public class MutationType
     /// </summary>
     /// <param name="options">Options for the schema builder</param>
     /// <typeparam name="TType"></typeparam>
-    public MutationType AddFrom<TType>(SchemaBuilderMutationOptions? options = null) where TType : class
+    public MutationType AddFrom<TType>(SchemaBuilderMethodOptions? options = null) where TType : class
     {
-        options ??= new SchemaBuilderMutationOptions();
+        options ??= new SchemaBuilderMethodOptions();
         var types = typeof(TType).Assembly
                                 .GetTypes()
                                 .Where(x => x.IsClass && !x.IsAbstract)
@@ -56,7 +56,7 @@ public class MutationType
     /// </summary>
     /// <param name="mutationDelegate">A method to execute the mutation logic</param>
     /// <param name="options">Options for the schema builder</param>
-    public MutationField Add(Delegate mutationDelegate, SchemaBuilderMutationOptions? options = null)
+    public MutationField Add(Delegate mutationDelegate, SchemaBuilderMethodOptions? options = null)
     {
         return Add(SchemaType.Schema.SchemaFieldNamer(mutationDelegate.Method.Name), mutationDelegate, options);
     }
@@ -67,7 +67,7 @@ public class MutationType
     /// <param name="mutationName">Mutation field name</param>
     /// <param name="mutationDelegate">A method to execute the mutation logic</param>
     /// <param name="options">Options for the schema builder</param>
-    public MutationField Add(string mutationName, Delegate mutationDelegate, SchemaBuilderMutationOptions? options = null)
+    public MutationField Add(string mutationName, Delegate mutationDelegate, SchemaBuilderMethodOptions? options = null)
     {
         var description = (mutationDelegate.Method.GetCustomAttribute(typeof(DescriptionAttribute)) as DescriptionAttribute)?.Description ?? string.Empty;
         return Add(mutationName, description, mutationDelegate, options);
@@ -80,14 +80,14 @@ public class MutationType
     /// <param name="description">Description of the mutation field</param>
     /// <param name="mutationDelegate">A method to execute the mutation logic</param>
     /// <param name="options">Options for the schema builder</param>
-    public MutationField Add(string mutationName, string description, Delegate mutationDelegate, SchemaBuilderMutationOptions? options = null)
+    public MutationField Add(string mutationName, string description, Delegate mutationDelegate, SchemaBuilderMethodOptions? options = null)
     {
         return AddMutationMethod(mutationName, null, mutationDelegate.Method, description, options);
     }
 
-    private MutationField AddMutationMethod(string name, RequiredAuthorization? classLevelRequiredAuth, MethodInfo method, string? description, SchemaBuilderMutationOptions? options)
+    private MutationField AddMutationMethod(string name, RequiredAuthorization? classLevelRequiredAuth, MethodInfo method, string? description, SchemaBuilderMethodOptions? options)
     {
-        options ??= new SchemaBuilderMutationOptions();
+        options ??= new SchemaBuilderMethodOptions();
         var isAsync = method.GetCustomAttribute(typeof(AsyncStateMachineAttribute)) != null;
         var takeGenericArgument = isAsync || method.ReturnType.IsGenericType && method.ReturnType.GetGenericTypeDefinition() == typeof(Task<>);
         var methodAuth = SchemaType.Schema.AuthorizationService.GetRequiredAuthFromMember(method);