+++ src/EntityGraphQL/Subscriptions/GraphQLSubscribeResult.cs
@@ -7,19 +7,20 @@ namespace EntityGraphQL.Subscriptions
     public class GraphQLSubscribeResult
     {
         public Type EventType { get; }
-        /// <summary>
-        /// Will be the IObservable<TQueryType>
-        /// </summary>
-        public object SubscriptionObservable { get; }
+
+        private readonly object observable;
+
         public GraphQLSubscriptionStatement SubscriptionStatement { get; }
         public GraphQLSubscriptionField Field { get; }
 
         public GraphQLSubscribeResult(Type eventType, object result, GraphQLSubscriptionStatement graphQLSubscriptionStatement, GraphQLSubscriptionField node)
         {
             EventType = eventType;
-            SubscriptionObservable = result;
+            observable = result;
             SubscriptionStatement = graphQLSubscriptionStatement;
             Field = node;
         }
+
+        public object GetObservable() => observable;
     }
 }