+++ CHANGELOG.md
@@ -15,7 +15,7 @@ services.AddGraphQLSchema<DemoContext>(options =>
 });
 ```
 
-- `AddMutationsFrom` and friends now take an optional `SchemaBuilderMutationOptions` options object to configure how mutations are built, following the `SchemaBuilderOptions` used elsewhere. See updated docs. Important defaults:
+- `AddMutationsFrom` and friends now take an optional `SchemaBuilderMethodOptions` options object to configure how mutations are built, following the `SchemaBuilderOptions` used elsewhere. See updated docs. Important defaults:
 
 ```c#
 bool AutoCreateInputTypes = true; // Any input types seen will be added to the schema
@@ -24,6 +24,7 @@ bool AutoCreateNewComplexTypes = true; // Return types of mutations will be adde
 ```
 
 - `SchemaBuilderOptions.IgnoreTypes` Now uses `Type` instead of `string`. It is a `HashSet<Type>` now to avoid confusion ofwhich name to use (full name space or not)
+- `MutationArgumentsAttribute` renamed to `GraphQLArgumentsAttribute` and is used with subscriptions or mutaiton method arguments
 
 ## Changes